hardware/requirements.txt
pyserial>=3.5
python-socketio[client]>=5.10.0
requests>=2.31.0
python-dotenv>=1.0.0

# Raspberry Pi only (comment out on Windows)
# RPi.GPIO>=0.7.1
